HealthBars
Fixed the Heatlhbar  for the Samurai then add a duplicate of it to the Werewolf.

Also added that damage is taken when the samurai is attacking and will take damage if hit by the werewolf.

A future update would be to use two different box colliders, one for the weapon and one for the body so that the samurai only takes damage when the werewolf hits it on the body